requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
unstructured==0.9.4
pandas==2.2.1
langchain==0.0.208
langchain-community==0.1.10
asyncpg==0.27.0
python-dotenv==1.0.0
langchain-unstructured
unstructured[docx]

